changed distance_calc return type to float

+++ Second_Milestone/GPS_Functions/functions.h
@@ -8,7 +8,7 @@
 #define PI 3.14159265358979323846
 #define R 6378137
 
-int distance_calc (float lat1, float lat2, float lng1, float lng2);
+float distance_calc (float lat1, float lat2, float lng1, float lng2);
 float nmea_to_map(float nmea);
 uint8_t Convert_fun(char str [],char N[],char S[]);
 void gps_read(char [100]);
